Prototype an AbstractTransformingClassLoader
It's a ClassLoader which gets resources from its backend ClassLoader
(different from parent) and defines them into classes as if they were
his own.
It can use this as an opportunity to make any kind of inspections or
modifications to the bytecode of the loaded classes.

It's a WIP - still gotta implement findResource(), add some logging, and
possibly improve the codesource guessing sorcery.