Use AnalyzerAdapter to hunt down the uninitialized references
In SandboxAdapter, when rewriting a constructor call, we need to change
its calling convention to that of a factory method. Regular constructor
calls first execute NEW, which creates an uninitialized object, and
later INVOKESPECIALs <init> on it. When we rewrite it to INVOKEDYNAMIC,
it no longer needs the uninitialized object passed as 0th argument.
We could either:
  a) remove NEW, but at this point we don't know if we want to rewrite,
     cause NEW has only class name, and not constructor signature; or
  b) replace all the references to the uninitialized object created by
     NEW with the reference returned by INVOKEDYNAMIC

This commit changes it so that we go the (b) route.

However, the implementation is still incomplite - it doesn't handle the
cases where an uninitialized object reference is buried deep in the
stack.