Add code to replace uninitialized refs deep in the stack
However inefficient that sounds, if there is a reference to our
uninitialized object buried deep in the stack, move stuff from stack to
locals until the uninitialized object ref is on top, then replace it with
the ref returned from INVOKEDYNAMIC, and put all the stuff from locals back
on stack.

This whole operation runs only when needed, which, fortunately, should
happen very rarely, as regular Java compilers don't put uninitialized refs
in such places that we'd need to dig for them.

Unfortunately, this also means this code won't be tested very often.