Replace NEW with ACONST_NULL if we know we'll rewrite the constructor call
If, at the time of encountering a NEW instruction, the policy is sure that
it wants to rewrite this constructor call, no matter what the argument
types might be, then we can replace the NEW instruction with ACONST_NULL,
putting null in place of the uninitialized object reference.

This will hopefully save us some overhead, and after rewriting the
INVOKESPECIAL, we will replace this null (in all the places where it
propagated) with the reference returned by INVOKEDYNAMIC, anyway.