Add a separate InvocationType for superinitializer calls
- add InvocationType.INVOKESUPERINITSPECIAL with same opcodes as
  INVOKESIECIAL, but not registered in opcode lookup maps
- make SandboxAdapter use INVOKESUPERINITSPECIAL instead of
  INVOKENEWSPECIAL to represent superinitializer calls

This will make it possible for RewritePolicy to distinguish
superinitializer calls from regular constructor calls. It will also
(hopefully) be useful in the future RuntimePolicy API for
superinitializers.

Note that INVOKESUPERINITSPECIAL.id() is the same as
INVOKESPECIAL's. As a result, turning INVOKESUPERINITSPECIAL to an
ID and back to an enum value will result in INVOKESPECIAL instead.
This shouldn't be a problem for now, as superinitializer calls have
a spearate bootstrap method. However, it may cause problems in the
future, especially library users use the IDs to serialize
InvocationTypes.