Implement rewrite of INVOKEDYNAMIC
Let the RewritePolicy change the bootstrap method and its
arguments.

We don't know which arguments of the bsm will be needed
at runtime. Moreover, some of them might be constant MethodHandles
which (depending on particular bsm) might need to be broken down
at rewrite time, so that at runtime we know what method they refer
to. We assume the RewritePolicy will know what bsm it's dealing
with, and will know all these details required to properly
intercept it.