Add default implementation of INVOKEDYNAMIC interception
Implement Bootstraps.wrapInvokeDynamic() which intercepts the
INVOKEDYNAMIC bootstrap process as if it was a regular call to
the bootstrap method.

Add AbstractRewritePolicy which rewrites INVOKEDYNAMICs so that
they are intercepted by Bootstraps.wrapInvokeDynamic(), inf
and only if the bootstrap method would also be intercepted
if it was called in a regular method call.

Make BlindPolicy and ChangeMindPolicy extend AbstractRewritePolicy