Split the JVM 8 dependent parts to a separate package
Varargs collection doesn't work for bootstrap methods on JVM 7
and earlier. This means our generic INVOKEDYNAMIC interception
mechanism requires JVM 8 features. Split it to a separate package,
and make Transformer and tests use it only when JVM 8 is detected.

This is a temporary solution, until we get a proper policy
composition framework.