Delete the old SandboxClassLoader
The old SandboxClassLoader was derived from OpenJDK's
URLClassLoader, which was kinda problematic, because it uses
sun.misc.* classes (not official API, OpenJDK/OracleJDK specific)
and is covered by GPLv2 with classpath exception, as opposed to
LGPLv3 which we're using.

Since we have a new classloader as a replacement,
remove the old one.